Fix the hystrix.command.default.metrics.rollingPercentile.enable=false not working after the command metrics refactoring
